refactor(main.js): move core functions to a dedicated file
