+++ src/main.ts
@@ -1,11 +1,16 @@
 import { Plugin } from 'obsidian';
 import { DEFAULT_SETTINGS, FileExplorerNoteCountSettings, FileExplorerNoteCountSettingsTab } from './settings';
 import './main.css'
+import { clearInsertedNumbers, updateFolderNumbers } from './core';
 
 export default class FileExplorerNoteCount extends Plugin {
 
 	settings: FileExplorerNoteCountSettings;
 	loadedStyles: Array<HTMLStyleElement>;
+
+	updateFolderNumbers = () => {
+    updateFolderNumbers(this.app);
+  };
 
 	async onload() {
 		this.addSettingTab(new FileExplorerNoteCountSettingsTab(this.app, this));
@@ -20,7 +25,7 @@ export default class FileExplorerNoteCount extends Plugin {
 	}
 
 	onunload() {
-		this.clearInsertedNumbers();
+		clearInsertedNumbers();
 		console.log('unloading plugin');
 		this.app.vault.off('create', this.updateFolderNumbers);
 		this.app.vault.off('rename', this.updateFolderNumbers);
@@ -34,62 +39,6 @@ export default class FileExplorerNoteCount extends Plugin {
 
 	async saveSettings() {
 		await this.saveData(this.settings);
-	}
-
-	updateFolderNumbers = () => {
-
-		this.clearInsertedNumbers();
-
-		// Get All Available Notes under Vault
-		var mdNotes = this.app.vault.getMarkdownFiles();
-
-		// Create Folder File Map
-		const counts: { [key: string]: number } = {};
-		mdNotes.forEach(mdNote => {
-			for (let folder = mdNote.parent; folder != null; folder = folder.parent) {
-				counts[folder.path] = 1 + (counts[folder.path] || 0)
-			}
-		})
-
-		// Loop Through File Explorer Elements
-		var fileExplorers = this.app.workspace.getLeavesOfType("file-explorer");
-
-		for (let fileExplorer of fileExplorers) {
-			// @ts-ignore
-			for (const [key, value] of Object.entries(fileExplorer.view.fileItems)) {
-
-				if (value.titleEl.className === 'nav-folder-title') {
-					// Get the Title Node
-					var folderTitleNode: HTMLElement = value.titleEl;
-
-					// Get Path of Folder Title
-					var currentDataPath = folderTitleNode.getAttr('data-path');
-
-					// No number for the Vault Main Folder
-					if (currentDataPath === '/') continue;
-
-					// Add Number of Notes
-					if (counts[currentDataPath]) {
-						var folderCount = folderTitleNode.createDiv({ cls: 'oz-folder-numbers', text: counts[currentDataPath].toString() })
-
-						// Add Attribute if There is A Children Folder for CSS
-						for (let child of value.children) {
-							if (child.children) {
-								folderCount.setAttr('haschild', 'true');
-								break;
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	clearInsertedNumbers = () => {
-		const allInsertedNumbers = document.querySelectorAll('.oz-folder-numbers');
-		allInsertedNumbers.forEach(insertedNumber => {
-			insertedNumber.parentNode.removeChild(insertedNumber);
-		})
 	}
 
 	// Style Settings
